This Asp.Net website helps people study English
of a website used to study English. Users can create a personal
dictionaries and used them to create practices tests such as puzzles,
fill-in-the-blanks and multiple choice. User can also merge their
dictionary to the public dictionary and their friends’ dictionary to
make it larger. This function helps people study in groups and
communicate with others.